More receive & resample tweaks
Do a blocking channel recv into the buffer if the non-blocking one fails
and the buffer is insufficiently full, instead of a sleep. This is more
efficient.

Rewrite some parts of the resampler for more clarity

Tweak the resampler for stability

This probably isn't the last change ever, it's got a bit of a ring in
the impulse response, but it's quite nice steady-state.