It's a real PID now!
The latency measurements now go through a filter which does some
smoothing and computes a slope, and the slope term is used in the
control loop. This allows making the loop response quite a bit faster
(and gives nice frequency convergence on startup) without causing
problems with overshoot/undershoot.

Still subject to change as usual.