Swap the "source" and "sink" options.
They *kinda* made sense with module-null-sink where nDAX was seeing the
opposite sides of the loopbacks as the user app, and the options were
named after what nDAX used, but with pipes there's only one device each
way and it's exactly the opposite of what we're calling it. Reduce the
confusion before it's too late.