Add an option to consume our own RX stream
There's a known problem with pipewire's implementation of the pulseaudio
pipe_source module that causes high RX latency if something isn't
listening on the RX audio device all the time (for instance, wsjtx might
work fine until the first time you transmit, then stop being able to
receive, because it shut down the RX device while it was transmitting,
and when it opened it again there would be a backlog of several seconds
of audio).

I've been recommending that people with pipewire open a mixer app like
pavucontrol because the level monitor would make sure that something was
always reading from the device, but it should work pretty much as well
to do it from nDAX directly.

This commit adds a new option `-consume`, with values `false`, `true`,
and `auto`. Setting it to `true` makes nDAX start reading from the RX
audio device as soon as it's created, and discard the data. The default,
`auto`, enables the behavior only if the PulseAudio "package_name"
string contains "PipeWire".